git-svn-id: svn+ssh://svn.pointclouds.org/pcl/branches/pcl-1.x@4681 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ common/include/pcl/common/time.h
@@ -68,7 +68,7 @@ namespace pcl
       getTime ()
       {
         boost::posix_time::ptime end_time = boost::posix_time::microsec_clock::local_time ();
-        return ((end_time - start_time_).total_milliseconds ());
+        return ((double) ((end_time - start_time_).total_milliseconds ()));
       }
 
       /** \brief Retrieve the time in seconds spent since the last call to \a reset(). */
