+++ octree/include/pcl/octree/impl/octree_base.hpp
@@ -305,7 +305,7 @@ namespace pcl
         deserializeTreeRecursive (binaryTreeVectorIterator, rootNode_, depthMask_, newKey, dataVectorIterator,
                                   dataVectorEndIterator);
 
-        objectCount_ = dataVector_arg.size();
+        objectCount_ = (unsigned int)dataVector_arg.size();
       }
 
     //////////////////////////////////////////////////////////////////////////////////////////////
@@ -327,7 +327,7 @@ namespace pcl
         deserializeTreeAndOutputLeafDataRecursive (binaryTreeVectorIterator, rootNode_, depthMask_, newKey,
                                                    dataVector_arg);
 
-        objectCount_ = dataVector_arg.size();
+        objectCount_ = (unsigned int)dataVector_arg.size();
       }
 
     //////////////////////////////////////////////////////////////////////////////////////////////
